create a ggplot from the output of summariseLargeScaleCharacteristics.

`r lifecycle::badge("experimental")`

Usage

plotComparedLargeScaleCharacteristics(
  data,
  referenceGroupLevel = NULL,
  referenceStrataLevel = NULL,
  referenceVariableLevel = NULL,
  referenceCdmName = NULL,
  splitStrata = FALSE,
  facet = NULL,
  colorVars = NULL,
  missings = 0
)

Arguments

data

output of summariseLargeScaleCharacteristics().

referenceGroupLevel

group_level value to be used as the reference.

referenceStrataLevel

strata_level value to be used as the reference.

referenceVariableLevel

variable_level value to be used as the reference.

referenceCdmName

cdm_name value to be used as the reference.

splitStrata

boolean variable (TRUE/FALSE)

facet

columns in data to facet. If the facet position wants to be specified, use the formula class for the input (e.g., strata + table_name ~ group_level + cdm_name). Variables before "~" will be facet by on horizontal axis, whereas those after "~" on vertical axis. Character format is also allowed (e.g., c("strata","table_name","group_level","cdm_name")). Only the following columns are allowed to be facet by: c("cdm_name", "group_level", "strata_level", "variable_level", "strata", "table_name"). If splitStrata = TRUE, strata levels are also allowed.

colorVars

column in data to color by. Only the following columns are allowed to be used: c("cdm_name", "group_level", "strata_level", "variable_level", "strata", "table_name"). If splitStrata = TRUE, strata levels are also allowed.

missings

value to replace the missings with.

Value

A ggplot.