Format a summarised_characteristics object into a visual table.

`r lifecycle::badge("experimental")`

Usage

tableCohortCount(
  result,
  type = "gt",
  formatEstimateName = c(N = "<count>"),
  header = c("group"),
  split = c("group", "strata"),
  groupColumn = NULL,
  excludeColumns = c("result_id", "estimate_type", "variable_level", "additional_name",
    "additional_level"),
  .options = list()
)

Arguments

result

A summarised_characteristics object.

type

Type of desired formatted table, possibilities: "gt", "flextable", "tibble".

formatEstimateName

Named list of estimate name's to join, sorted by computation order. Indicate estimate_name's between <...>.

header

A vector containing which elements should go into the header in order. Allowed are: `cdm_name`, `group`, `strata`, `additional`, `variable`, `estimate`, `settings`.

split

A vector containing the name-level groups to split ("group", "strata", "additional"), or an empty character vector to not split.

groupColumn

Column to use as group labels.

excludeColumns

Columns to drop from the output table.

.options

Named list with additional formatting options. CohortCharacteristics::optionsTableCharacteristics() shows allowed arguments and their default values.

Value

A table with a formatted version of the summariseCohortCount result result.

Examples

# \donttest{
library(CohortCharacteristics)

cdm <- mockCohortCharacteristics()

cdm$cohort1 |>
  summariseCharacteristics() |>
  tableCharacteristics()
#> ℹ adding demographics columns
#> ℹ summarising data
#> ✔ summariseCharacteristics finished!
#> ! Results have not been suppressed.

  

CDM name
      	Variable name
      	Variable level
      	Estimate name
      	
        Cohort name
      
    

Cohort 1
      	Cohort 2
      	Cohort 3
    

PP_MOCK
	Number records
	-
	N
	5
	3
	2

	Number subjects
	-
	N
	5
	3
	2

	Cohort start date
	-
	Median [Q25 - Q75]
	2020-06-27 [1956-05-03 - 2044-06-08]
	1960-03-26 [1959-04-28 - 1983-11-05]
	1999-11-04 [1978-05-23 - 2021-04-17]

	
	
	Range
	1939-12-28 to 2132-05-15
	1958-05-30 to 2007-06-16
	1956-12-10 to 2042-09-28

	Cohort end date
	-
	Median [Q25 - Q75]
	2062-02-02 [2053-09-14 - 2092-12-01]
	2008-04-03 [1998-09-21 - 2011-01-08]
	2019-08-24 [1988-04-30 - 2050-12-17]

	
	
	Range
	1984-01-15 to 2136-05-30
	1989-03-11 to 2013-10-14
	1957-01-04 to 2082-04-12

	Sex
	Female
	N (%)
	4 (80.0%)
	1 (33.3%)
	2 (100.0%)

	
	Male
	N (%)
	1 (20.0%)
	2 (66.7%)
	-

	Age
	-
	Median [Q25 - Q75]
	56 [18 - 90]
	45 [40 - 50]
	16 [10 - 23]

	
	
	Mean (SD)
	59.00 (43.69)
	44.33 (10.02)
	16.50 (17.68)

	
	
	Range
	16 to 115
	34 to 54
	4 to 29

	Prior observation
	-
	Median [Q25 - Q75]
	20,632 [6,935 - 33,031]
	16,585 [14,544 - 18,237]
	6,334 [4,069 - 8,598]

	
	
	Mean (SD)
	21,740.40 (15,911.69)
	16,325.67 (3,699.82)
	6,333.50 (6,404.27)

	
	
	Range
	5,967 to 42,137
	12,503 to 19,889
	1,805 to 10,862

	Future observation
	-
	Median [Q25 - Q75]
	10,295 [7,493 - 22,095]
	13,092 [7,010 - 16,526]
	10,447 [5,338 - 15,556]

	
	
	Mean (SD)
	19,779.60 (21,929.82)
	11,326.67 (9,639.01)
	10,447.00 (14,449.02)

	
	
	Range
	2,230 to 56,785
	927 to 19,961
	230 to 20,664

mockDisconnect(cdm = cdm)
# }