Format a summarise_characteristics object into a visual table.

`r lifecycle::badge("experimental")`

Usage

tableCharacteristics(
  result,
  type = "gt",
  formatEstimateName = c(`N (%)` = "<count> (<percentage>%)", N = "<count>",
    `Median [Q25 - Q75]` = "<median> [<q25> - <q75>]", `Mean (SD)` = "<mean> (<sd>)",
    Range = "<min> to <max>"),
  header = c("group"),
  split = c("group", "strata"),
  groupColumn = NULL,
  excludeColumns = c("result_id", "estimate_type", "additional_name", "additional_level"),
  .options = list()
)

Arguments

result

A summarise_characteristics object.

type

Type of desired formatted table, possibilities: "gt", "flextable", "tibble".

formatEstimateName

Named list of estimate name's to join, sorted by computation order. Indicate estimate_name's between <...>.

header

A vector containing which elements should go into the header in order. Allowed are: `cdm_name`, `group`, `strata`, `additional`, `variable`, `estimate`, `settings`.

split

A vector containing the name-level groups to split ("group", "strata", "additional"), or an empty character vector to not split.

groupColumn

Column to use as group labels.

excludeColumns

Columns to drop from the output table.

.options

Named list with additional formatting options. CohortCharacteristics::optionsTableCharacteristics() shows allowed arguments and their default values.

Value

A table with a formatted version of the summariseCharacteristics result.

Examples

# \donttest{
library(CohortCharacteristics)

cdm <- mockCohortCharacteristics()
#> ! cohort columns will be reordered to match the expected order:
#>   cohort_definition_id, subject_id, cohort_start_date, and cohort_end_date.
#> ! cohort columns will be reordered to match the expected order:
#>   cohort_definition_id, subject_id, cohort_start_date, and cohort_end_date.

cdm$cohort1 |>
  summariseCharacteristics() |>
  tableCharacteristics()
#> ℹ adding demographics columns
#> ℹ summarising data
#> ✔ summariseCharacteristics finished!
#> ! Results have not been suppressed.

  

CDM name
      	Variable name
      	Variable level
      	Estimate name
      	
        Cohort name
      
    

Cohort 2
      	Cohort 3
      	Cohort 1
    

PP_MOCK
	Number records
	-
	N
	4
	4
	2

	Number subjects
	-
	N
	4
	4
	2

	Cohort start date
	-
	Median [Q25 - Q75]
	2045-04-14 [2039-09-14 - 2073-01-12]
	2023-03-30 [1950-04-27 - 2091-04-30]
	2020-12-21 [1980-10-21 - 2061-02-20]

	
	
	Range
	2036-06-23 to 2142-10-06
	1931-11-23 to 2095-03-30
	1940-08-21 to 2101-04-23

	Cohort end date
	-
	Median [Q25 - Q75]
	2057-05-18 [2052-07-10 - 2082-05-25]
	2040-01-12 [1985-01-24 - 2092-09-11]
	2061-10-26 [2037-10-23 - 2085-10-30]

	
	
	Range
	2050-04-29 to 2145-01-30
	1975-08-12 to 2095-03-31
	2013-10-20 to 2109-11-03

	Sex
	Female
	N (%)
	1 (25.0%)
	3 (75.0%)
	2 (100.0%)

	
	Male
	N (%)
	3 (75.0%)
	1 (25.0%)
	-

	Age
	-
	Median [Q25 - Q75]
	71 [54 - 104]
	62 [24 - 95]
	64 [34 - 94]

	
	
	Mean (SD)
	87.75 (56.95)
	57.75 (45.68)
	64.00 (84.85)

	
	
	Range
	40 to 169
	8 to 99
	4 to 124

	Prior observation
	-
	Median [Q25 - Q75]
	26,164 [20,102 - 38,368]
	22,773 [9,156 - 34,858]
	23,548 [12,621 - 34,475]

	
	
	Mean (SD)
	32,306.25 (20,800.88)
	21,240.25 (16,558.78)
	23,548.00 (30,906.22)

	
	
	Range
	14,894 to 62,004
	3,248 to 36,167
	1,694 to 45,402

	Future observation
	-
	Median [Q25 - Q75]
	22,814 [17,872 - 31,551]
	10,913 [1,176 - 26,791]
	25,866 [19,788 - 31,944]

	
	
	Mean (SD)
	26,608.75 (17,529.17)
	17,054.00 (21,611.63)
	25,866.00 (17,191.18)

	
	
	Range
	9,737 to 51,071
	4 to 46,386
	13,710 to 38,022

mockDisconnect(cdm = cdm)
# }