Format a summarise_characteristics object into a visual table.

`r lifecycle::badge("experimental")`

Usage

tableCharacteristics(
  result,
  type = "gt",
  formatEstimateName = c(`N (%)` = "<count> (<percentage>%)", N = "<count>",
    `Median [Q25 - Q75]` = "<median> [<q25> - <q75>]", `Mean (SD)` = "<mean> (<sd>)",
    Range = "<min> to <max>"),
  header = c("group"),
  split = c("group", "strata"),
  groupColumn = NULL,
  excludeColumns = c("result_id", "estimate_type", "additional_name", "additional_level"),
  .options = list()
)

Arguments

result

A summarise_characteristics object.

type

Type of desired formatted table, possibilities: "gt", "flextable", "tibble".

formatEstimateName

Named list of estimate name's to join, sorted by computation order. Indicate estimate_name's between <...>.

header

A vector containing which elements should go into the header in order. Allowed are: `cdm_name`, `group`, `strata`, `additional`, `variable`, `estimate`, `settings`.

split

A vector containing the name-level groups to split ("group", "strata", "additional"), or an empty character vector to not split.

groupColumn

Column to use as group labels.

excludeColumns

Columns to drop from the output table.

.options

Named list with additional formatting options. CohortCharacteristics::optionsTableCharacteristics() shows allowed arguments and their default values.

Value

A table with a formatted version of the summariseCharacteristics result.

Examples

# \donttest{
library(CohortCharacteristics)

cdm <- mockCohortCharacteristics()

cdm$cohort1 |>
  summariseCharacteristics() |>
  tableCharacteristics()
#> ℹ adding demographics columns
#> ℹ summarising data
#> ✔ summariseCharacteristics finished!
#> Warning: The `split` argument of `visOmopTable()` is deprecated as of visOmopResults
#> 0.4.0.
#> ℹ The deprecated feature was likely used in the CohortCharacteristics package.
#>   Please report the issue at
#>   <https://github.com/darwin-eu-dev/CohortCharacteristics/issues>.
#> ! Results have not been suppressed.

  

CDM name
      	Variable name
      	Variable level
      	Estimate name
      	
        Cohort name
      
    

cohort_3
      	cohort_1
      	cohort_2
    

PP_MOCK
	Number records
	-
	N
	4
	4
	2

	Number subjects
	-
	N
	4
	4
	2

	Cohort start date
	-
	Median [Q25 - Q75]
	1924-03-17 [1912-01-20 - 1938-07-16]
	1931-09-20 [1924-07-22 - 1942-07-17]
	1932-06-21 [1928-08-01 - 1936-05-11]

	
	
	Range
	1909-02-13 to 1947-12-30
	1916-01-10 to 1962-01-22
	1924-09-12 to 1940-03-30

	Cohort end date
	-
	Median [Q25 - Q75]
	1939-02-01 [1922-04-07 - 1957-05-28]
	1932-08-23 [1926-04-19 - 1946-03-13]
	1939-04-10 [1936-12-15 - 1941-08-03]

	
	
	Range
	1914-12-27 to 1969-03-06
	1919-03-25 to 1974-11-21
	1934-08-22 to 1943-11-27

	Sex
	Female
	N (%)
	1 (25.0%)
	2 (50.0%)
	-

	
	Male
	N (%)
	3 (75.0%)
	2 (50.0%)
	2 (100.0%)

	Age
	-
	Median [Q25 - Q75]
	9 [5 - 14]
	14 [7 - 20]
	22 [20 - 24]

	
	
	Mean (SD)
	10.25 (6.40)
	13.25 (9.43)
	22.00 (5.66)

	
	
	Range
	5 to 18
	2 to 22
	18 to 26

	Prior observation
	-
	Median [Q25 - Q75]
	3,380 [1,862 - 5,401]
	5,313 [2,666 - 7,534]
	8,207 [7,518 - 8,896]

	
	
	Mean (SD)
	3,883.25 (2,488.46)
	4,886.50 (3,486.13)
	8,207.00 (1,948.79)

	
	
	Range
	1,837 to 6,937
	739 to 8,181
	6,829 to 9,585

	Future observation
	-
	Median [Q25 - Q75]
	10,982 [8,634 - 13,695]
	8,524 [6,552 - 9,041]
	3,920 [2,746 - 5,094]

	
	
	Mean (SD)
	11,346.50 (3,422.98)
	7,068.00 (4,334.37)
	3,920.00 (3,321.99)

	
	
	Range
	8,177 to 15,244
	719 to 10,504
	1,571 to 6,269

	Days in cohort
	-
	Median [Q25 - Q75]
	6,149 [4,445 - 6,892]
	815 [399 - 2,050]
	2,485 [1,912 - 3,058]

	
	
	Mean (SD)
	5,188.00 (3,096.95)
	1,634.00 (2,075.09)
	2,485.00 (1,622.10)

	
	
	Range
	716 to 7,738
	219 to 4,687
	1,338 to 3,632

mockDisconnect(cdm = cdm)
# }