Format a summarised_characteristics object into a visual table.

`r lifecycle::badge("experimental")`

Usage

tableCharacteristics(
  result,
  type = "gt",
  formatEstimateName = c(`N (%)` = "<count> (<percentage>%)", N = "<count>",
    `Median [Q25 - Q75]` = "<median> [<q25> - <q75>]", `Mean (SD)` = "<mean> (<sd>)",
    Range = "<min> to <max>"),
  header = c("group"),
  split = c("group", "strata"),
  groupColumn = NULL,
  excludeColumns = c("result_id", "estimate_type", "additional_name", "additional_level"),
  .options = list()
)

Arguments

result

A summarised_characteristics object.

type

Type of desired formatted table, possibilities: "gt", "flextable", "tibble".

formatEstimateName

Named list of estimate name's to join, sorted by computation order. Indicate estimate_name's between <...>.

header

A vector containing which elements should go into the header in order. Allowed are: `cdm_name`, `group`, `strata`, `additional`, `variable`, `estimate`, `settings`.

split

A vector containing the name-level groups to split ("group", "strata", "additional"), or an empty character vector to not split.

groupColumn

Column to use as group labels.

excludeColumns

Columns to drop from the output table.

.options

Named list with additional formatting options. CohortCharacteristics::optionsTableCharacteristics() shows allowed arguments and their default values.

Value

A table with a formatted version of the summariseCharacteristics result.

Examples

# \donttest{
library(CohortCharacteristics)

cdm <- mockCohortCharacteristics()

cdm$cohort1 |>
  summariseCharacteristics() |>
  tableCharacteristics()
#> ℹ adding demographics columns
#> ℹ summarising data
#> ✔ summariseCharacteristics finished!
#> ! Results have not been suppressed.

  

CDM name
      	Variable name
      	Variable level
      	Estimate name
      	
        Cohort name
      
    

Cohort 2
      	Cohort 1
      	Cohort 3
    

PP_MOCK
	Number records
	-
	N
	3
	5
	2

	Number subjects
	-
	N
	3
	5
	2

	Cohort start date
	-
	Median [Q25 - Q75]
	1996-07-02 [1984-03-06 - 2022-05-28]
	2018-11-02 [2017-07-27 - 2048-08-17]
	2037-02-22 [2028-02-16 - 2046-03-02]

	
	
	Range
	1971-11-08 to 2048-04-22
	1970-06-20 to 2107-04-16
	2019-02-09 to 2055-03-09

	Cohort end date
	-
	Median [Q25 - Q75]
	2064-02-29 [2043-06-15 - 2102-08-02]
	2026-07-09 [2019-03-03 - 2065-10-08]
	2050-11-11 [2035-02-25 - 2066-07-28]

	
	
	Range
	2022-09-28 to 2141-01-02
	1985-05-30 to 2118-02-05
	2019-06-12 to 2082-04-12

	Sex
	Female
	N (%)
	2 (66.7%)
	3 (60.0%)
	2 (100.0%)

	
	Male
	N (%)
	1 (33.3%)
	2 (40.0%)
	-

	Age
	-
	Median [Q25 - Q75]
	25 [16 - 28]
	57 [38 - 69]
	89 [86 - 92]

	
	
	Mean (SD)
	21.33 (12.90)
	57.60 (39.95)
	89.00 (9.90)

	
	
	Range
	7 to 32
	8 to 116
	82 to 96

	Prior observation
	-
	Median [Q25 - Q75]
	9,442 [6,090 - 10,621]
	21,027 [14,108 - 25,372]
	32,560 [31,274 - 33,846]

	
	
	Mean (SD)
	7,993.67 (4,700.92)
	21,241.40 (14,519.75)
	32,560.00 (3,635.94)

	
	
	Range
	2,739 to 11,800
	3,227 to 42,473
	29,989 to 35,131

	Future observation
	-
	Median [Q25 - Q75]
	22,226 [15,321 - 42,273]
	10,301 [7,892 - 11,991]
	9,688 [5,246 - 14,129]

	
	
	Mean (SD)
	30,987.33 (27,999.66)
	12,666.40 (9,779.76)
	9,687.50 (12,563.17)

	
	
	Range
	8,416 to 62,320
	3,862 to 29,286
	804 to 18,571

mockDisconnect(cdm = cdm)
# }