Format a summarise_characteristics object into a visual table.

`r lifecycle::badge("experimental")`

Usage

tableCharacteristics(
  result,
  type = "gt",
  header = c("cdm_name", "cohort_name"),
  groupColumn = NULL
)

Arguments

result

A summarise_characteristics object.

type

Type of desired formatted table, possibilities: "gt", "flextable", "tibble".

header

A vector containing which elements should go into the header in order. Allowed are: `cdm_name`, `group`, `strata`, `additional`, `variable`, `estimate`, `settings`.

groupColumn

Column to use as group labels.

Value

A table with a formatted version of the summariseCharacteristics result.

Examples

# \donttest{
cdm <- mockCohortCharacteristics()

result <- summariseCharacteristics(cdm$cohort1)
#> ℹ adding demographics columns
#> ℹ summarising data
#> ✔ summariseCharacteristics finished!

tableCharacteristics(result)
#> ! Results have not been suppressed.

  

      	
        CDM name
      
    

      	
        PP_MOCK
      
    

Variable name
      	Variable level
      	Estimate name
      	
        Cohort name
      
    

cohort_1
      	cohort_2
      	cohort_3
    

Number records
	-
	N
	5
	1
	4

Number subjects
	-
	N
	5
	1
	4

Cohort start date
	-
	Median [Q25 - Q75]
	1939-12-29 [1938-02-22 - 1962-10-02]
	1938-10-11 [1938-10-11 - 1938-10-11]
	1976-11-04 [1966-05-26 - 1985-09-27]

	
	Range
	1901-07-04 to 1983-05-23
	1938-10-11 to 1938-10-11
	1957-06-13 to 1990-01-17

Cohort end date
	-
	Median [Q25 - Q75]
	1961-03-28 [1953-04-28 - 1970-07-13]
	1952-03-26 [1952-03-26 - 1952-03-26]
	1981-12-29 [1972-04-13 - 1988-06-04]

	
	Range
	1907-07-01 to 1988-09-02
	1952-03-26 to 1952-03-26
	1960-07-07 to 1990-05-10

Age
	-
	Median [Q25 - Q75]
	7 [3 - 12]
	18 [18 - 18]
	18 [14 - 18]

	
	Mean (SD)
	10.20 (11.43)
	-
	14.50 (7.68)

	
	Range
	0 to 29
	18 to 18
	3 to 19

Sex
	Female
	N (%)
	1 (20.00%)
	1 (100.00%)
	3 (75.00%)

	Male
	N (%)
	4 (80.00%)
	-
	1 (25.00%)

Prior observation
	-
	Median [Q25 - Q75]
	2,831 [1,148 - 4,745]
	6,858 [6,858 - 6,858]
	6,664 [5,252 - 6,816]

	
	Mean (SD)
	3,928.40 (4,181.46)
	-
	5,403.50 (2,785.58)

	
	Range
	184 to 10,734
	6,858 to 6,858
	1,235 to 7,050

Future observation
	-
	Median [Q25 - Q75]
	7,153 [6,657 - 7,427]
	5,841 [5,841 - 5,841]
	6,181 [2,774 - 9,616]

	
	Mean (SD)
	7,748.20 (3,935.79)
	-
	6,209.25 (4,407.52)

	
	Range
	3,353 to 14,151
	5,841 to 5,841
	1,808 to 10,667

Days in cohort
	-
	Median [Q25 - Q75]
	2,842 [2,189 - 5,545]
	4,916 [4,916 - 4,916]
	1,196 [869 - 1,576]

	
	Mean (SD)
	4,053.40 (2,521.08)
	-
	1,249.75 (975.13)

	
	Range
	1,930 to 7,761
	4,916 to 4,916
	114 to 2,493

mockDisconnect(cdm = cdm)
# }