Format a summarise_characteristics object into a visual table.

`r lifecycle::badge("experimental")`

Usage

tableCohortCount(
  result,
  type = "gt",
  formatEstimateName = c(N = "<count>"),
  header = c("group"),
  split = c("group", "strata"),
  groupColumn = NULL,
  excludeColumns = c("result_id", "estimate_type", "variable_level", "additional_name",
    "additional_level"),
  .options = list()
)

Arguments

result

A summarise_characteristics object.

type

Type of desired formatted table, possibilities: "gt", "flextable", "tibble".

formatEstimateName

Named list of estimate name's to join, sorted by computation order. Indicate estimate_name's between <...>.

header

A vector containing which elements should go into the header in order. Allowed are: `cdm_name`, `group`, `strata`, `additional`, `variable`, `estimate`, `settings`.

split

A vector containing the name-level groups to split ("group", "strata", "additional"), or an empty character vector to not split.

groupColumn

Column to use as group labels.

excludeColumns

Columns to drop from the output table.

.options

Named list with additional formatting options. CohortCharacteristics::optionsTableCharacteristics() shows allowed arguments and their default values.

Value

A table with a formatted version of the summariseCohortCount result result.

Examples

# \donttest{
library(CohortCharacteristics)

cdm <- mockCohortCharacteristics()
#> ! cohort columns will be reordered to match the expected order:
#>   cohort_definition_id, subject_id, cohort_start_date, and cohort_end_date.
#> ! cohort columns will be reordered to match the expected order:
#>   cohort_definition_id, subject_id, cohort_start_date, and cohort_end_date.

cdm$cohort1 |>
  summariseCharacteristics() |>
  tableCharacteristics()
#> ℹ adding demographics columns
#> ℹ summarising data
#> ✔ summariseCharacteristics finished!
#> ! Results have not been suppressed.

  

CDM name
      	Variable name
      	Variable level
      	Estimate name
      	
        Cohort name
      
    

Cohort 3
      	Cohort 2
      	Cohort 1
    

PP_MOCK
	Number records
	-
	N
	6
	3
	1

	Number subjects
	-
	N
	6
	3
	1

	Cohort start date
	-
	Median [Q25 - Q75]
	2052-01-21 [2001-06-23 - 2080-05-18]
	2126-09-28 [2107-10-14 - 2141-07-24]
	2009-12-10 [2009-12-10 - 2009-12-10]

	
	
	Range
	1984-02-23 to 2130-04-08
	2088-10-29 to 2156-05-19
	2009-12-10 to 2009-12-10

	Cohort end date
	-
	Median [Q25 - Q75]
	2066-11-01 [2009-01-20 - 2098-07-07]
	2131-06-01 [2118-10-20 - 2145-08-24]
	2010-07-30 [2010-07-30 - 2010-07-30]

	
	
	Range
	1991-07-25 to 2136-11-29
	2106-03-10 to 2159-11-17
	2010-07-30 to 2010-07-30

	Sex
	Female
	N (%)
	1 (16.7%)
	2 (66.7%)
	-

	
	Male
	N (%)
	5 (83.3%)
	1 (33.3%)
	1 (100.0%)

	Age
	-
	Median [Q25 - Q75]
	46 [28 - 92]
	137 [118 - 146]
	24 [24 - 24]

	
	
	Mean (SD)
	67.50 (66.81)
	130.00 (28.16)
	-

	
	
	Range
	0 to 184
	99 to 154
	24 to 24

	Prior observation
	-
	Median [Q25 - Q75]
	17,187 [10,401 - 33,650]
	50,177 [43,319 - 53,348]
	9,109 [9,109 - 9,109]

	
	
	Mean (SD)
	24,834.17 (24,389.65)
	47,718.67 (10,251.99)
	-

	
	
	Range
	53 to 67,302
	36,461 to 56,518
	9,109 to 9,109

	Future observation
	-
	Median [Q25 - Q75]
	18,232 [8,164 - 20,466]
	5,203 [3,627 - 8,480]
	3,624 [3,624 - 3,624]

	
	
	Mean (SD)
	14,461.50 (8,144.66)
	6,336.67 (4,950.82)
	-

	
	
	Range
	3,013 to 21,273
	2,051 to 11,756
	3,624 to 3,624

mockDisconnect(cdm = cdm)
# }